fynegen: basic multiple return values
If a go function returns multiple values, its return values are returned
as a dict in rye. If the arguments are named, they are keyed with their
name. If they aren't named, the result names are "1", "2", ...; except if there
is an error result at the end, which will be named "err".